This is a cross compiler setup script for ubuntu
It is modified to use apt-get instead of dnf and also uses axel instead of wget for wayyy faster downloading, but if your computer isn't powerful enough then change the -n numbers to something like 10 or 20 on likes 14-16 or it might crash. This version also dosen't use sudo except for the apt-get commands because in Ubuntu the home folder is fully owned by the user that it belongs to and not partially to root.